glues_it - Sketchy Wordpress Plugin Function
An old client told me he hired someone else to make few updates/changes
to his Wordpress site. After a while he called me telling me that the
person who was currently working on the site couldn’t be able to
install a new plugin.

Wordpress was returning an error which said like this:

Fatal error: Cannot redeclare glues_it() (previously declared in

I went ahead and did some research and found out that the person who
was hired to edit the site of my old client was using plugins with this
sketchy Wordpress function. I believe that these are premium plugins
being pirated.

I’m still looking out what it really does but for the time being I told
the site owner to get rid of the plugins with the function.